fix: corrige logs duplicados em campanhas, HTTP 400 no MM e migra broadcasts para Redis
- ctype_digit() para distinguir IDs inteiros de UUIDs no SendWhatsAppMessageHandler,
  eliminando logs duplicados em envios de campanha
- fallback de templateName no payloadData quando content está vazio (HTTP 400 na VPS)
- resolveTokens() converte formato list [{label,value}] → chave-valor antes do buildPayload()
  no MarketingMessageSubscriber e WhatsAppMessageModel
- sendToLists() migrado de AMQP para Redis (whatsapp_direct) via SendWhatsAppDirectBatchMessage,
  isolando fila de MM da fila de campanhas
- BulkRateLimiter::getBulkSendDelay() para calcular intervalo entre mensagens a partir
  de bulk_rate_per_minute da config do plugin
- sendDelay alterado de int para float em SendWhatsAppDirectBatchMessage;
  usleep() recebe cast explícito para int
- testes de AMQP removidos; suite atualizada e validada (739 testes, 1359 assertions)

+++ Message/SendWhatsAppDirectBatchMessage.php
@@ -29,7 +29,7 @@ class SendWhatsAppDirectBatchMessage
     public function __construct(
         public readonly array $items,
         public readonly int $batchLimit,
-        public readonly int $sendDelay,
+        public readonly float $sendDelay,
     ) {
     }
 }
